Movement System Updates
The movement system has been cleaned up and is now mostly functional. I still need to come up with a way to consistently maintain the last direction entered if you continue to hold that direction down. At the moment, if you are holding a direction down and simultaneously press a direction that is tested for before that, the direction of higher precedence is taken. This means that if you held down left and then held down up on the very next tile, the game would move up and never left until you released up.

Also need to build collision detection, but that should be easy, as SaGa's collision system is based on the tilemap itself.