Layered Maps & Collision
Tile-based collision has been resolved by means of using a layered map system. There are now two background layers, a sprite layer, and a collision data "layer" that governs tile behavior. The collision layer is not truly a layer, just a table with the same dimensions as the tilemap it corresponds to.