Pseudocoded Elevation Handling
Am working on the outline for how handling two different elevations in the same map will work (going up a set of stairs to a higher area on the same map, for instance). This also includes handling bridges that the player either passes under whn on the lower plane or passes over when on the higher plane.

Additionally, outlined very briefly how scripting will work. It still needs more work, but the seeds have been planted.

Need to come back and add comments for quick reference in a lot of areas. Need to start compiling documentation before a lot of this evaporates from my mind.