Animated Tileset
Animated tileset capabilities have been added. To use, a sequential, unbroken set of animated tiles must be included in the map's tileset - one for each frame in the animation. Additionally, a few new fields have been added to the map metadata specification that indicate the animation loop's properties, such as the start and end tiles (and by virtue of that, the number of frames in the animation loop), and frame delay settings. Documentation to follow.